Updating Server side
Adding more functionality to server. Added functionality to count how many busses on the simulation are going to go to the stop the client is at. Added dictionary managers to efficiently store data obtained from sumo and the client  in a way that it can be easily used by the server thread as well as the main thread.