Updated App
Added functionality to read multiple data sent through from client. Able to receive all data for the given bus information, and display the data to the app. Able to receive data from the user input and store it in a string. Able to send the user input to the server. Added functionality to the refresh button.